+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-invalid-action-type-list.ts
@@ -1,4 +1,4 @@
-import { Action } from '../../action.decorator';
+import { Action } from '../../../../decorators/action.decorator';
 
 export class WithInvalidActionTypeList {
   @Action([undefined, null, ''] as any)
