+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-multiple-action-types.ts
@@ -1,4 +1,4 @@
-import { Action } from '../../action.decorator';
+import { Action } from '../../../../decorators/action.decorator';
 
 export class WithMultipleActionTypes {
   @Action(['0', '1'])
