+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-null-action-type.ts
@@ -1,4 +1,4 @@
-import { Action } from '../../action.decorator';
+import { Action } from '../../../../decorators/action.decorator';
 
 export class WithNullActionType {
   @Action(null as any)
