+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-undefined-action-type.ts
@@ -1,4 +1,4 @@
-import { Action } from '../../action.decorator';
+import { Action } from '../../../../decorators/action.decorator';
 
 export class WithUndefinedActionType {
   @Action(undefined as any)
